♻️ Refactered PlayerPage's components
