+++ src/components/SystemSettings/General/OwnerSetter/OwnerDialog.vue
@@ -5,7 +5,7 @@ import { Player } from 'app/src-electron/schema/player';
 import { usePlayerStore } from 'src/stores/WorldTabs/PlayerStore';
 import SsBtn from 'src/components/util/base/ssBtn.vue';
 import BaseDialogCard from 'src/components/util/baseDialog/baseDialogCard.vue';
-import SearchResultView from 'src/components/World/Player/SearchResultView.vue';
+import SearchResultCard from 'src/components/util/SearchResultCard.vue';
 import PlayerCard from '../PlayerCard.vue';
 import InputFieldView from './InputFieldView.vue';
 import { OwnerDialogProp, ReturnOwnerDialog } from './iOwnerDialog';
@@ -52,7 +52,7 @@ function registOwner() {
 
         <div v-show="playerStore.searchName !== ''" class="q-pb-md">
           <span class="text-caption">{{ $t('owner.searchResult') }}</span>
-          <SearchResultView
+          <SearchResultCard
             :register-btn-text="$t('owner.registerPlayer')"
             :register-process="ownerRegister"
           />
