+++ src/components/World/Player/CardView/core/GroupCard.vue
@@ -4,8 +4,8 @@ import { PlayerUUID } from 'app/src-electron/schema/brands';
 import { usePlayerStore } from 'src/stores/WorldTabs/PlayerStore';
 import SsBtn from 'src/components/util/base/ssBtn.vue';
 import SsTooltip from 'src/components/util/base/ssTooltip.vue';
-import BaseActionsCard from '../utils/BaseActionsCard.vue';
-import PlayerHeadView from './utils/PlayerHeadView.vue';
+import PlayerHeadAvatar from 'src/components/util/PlayerHeadAvatar.vue';
+import BaseActionsCard from 'src/components/World/utils/BaseActionsCard.vue';
 
 interface Prop {
   name: string;
@@ -58,7 +58,7 @@ async function onCardClicked() {
         <!-- TODO: 大量のプレイヤーが存在する（カードの高さが一定以上になる？）場合には折り畳みにすることを検討？ -->
         <div class="row q-gutter-md q-pt-sm">
           <template v-for="uuid in players" :key="uuid">
-            <PlayerHeadView :player="cachePlayers[uuid]" size="1.5rem" />
+            <PlayerHeadAvatar :player="cachePlayers[uuid]" size="1.5rem" />
           </template>
         </div>
       </q-card-section>
