+++ src/components/World/Player/CardView/core/OpMenu.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { OpLevel, OpSetting } from 'app/src-electron/schema/player';
-import { ServerProperties } from 'app/src-electron/schema/serverproperty';
 import { isValid } from 'src/scripts/error';
 import { assets } from 'src/assets/assets';
 import { useConsoleStore } from 'src/stores/ConsoleStore';
@@ -8,14 +7,24 @@ import { useMainStore } from 'src/stores/MainStore';
 import { usePlayerStore } from 'src/stores/WorldTabs/PlayerStore';
 import OpLevelBtn from 'src/components/World/Player/utils/OpLevelBtn.vue';
 
-interface Prop {
-  validProperties: ServerProperties;
-}
-defineProps<Prop>();
-
 const mainStore = useMainStore();
 const playerStore = usePlayerStore();
 const consoleStore = useConsoleStore();
+
+const isValidBtn = (opLevel: 0 | OpLevel) => {
+  // 権限無し or サーバー起動前なら設定可能
+  if (opLevel === 0 || consoleStore.status(mainStore.world.id) === 'Stop') {
+    return true;
+  }
+
+  // サーバー起動中は`op-permission-level`のLevelのみ設定可能
+  if (isValid(mainStore.world.properties)) {
+    return mainStore.world.properties['op-permission-level'] === opLevel;
+  }
+
+  // その他は設定不可
+  return false;
+};
 
 function setOP(setVal: 0 | OpLevel) {
   function setter(setVal?: OpSetting) {
@@ -70,10 +79,7 @@ function removePlayer() {
           :label="
             opLevel !== 0 ? $t('player.opLevel') + opLevel : $t('player.noOp')
           "
-          :disable="
-            consoleStore.status(mainStore.world.id) !== 'Stop' &&
-            ![validProperties['op-permission-level'], 0].includes(opLevel)
-          "
+          :disable="!isValidBtn(opLevel)"
           @click="() => setOP(opLevel)"
         />
       </template>
