+++ src/components/World/Player/CardView/core/parts/GroupBadgeView.vue
